Remove PdfLabel
