+++ app/src/main/java/com/books/hondana/Connection/KiiRequestConnection.java
@@ -3,7 +3,6 @@ package com.books.hondana.Connection;
 import android.support.annotation.Nullable;
 import android.util.Log;
 
-import com.books.hondana.Model.Book;
 import com.books.hondana.Model.Request;
 import com.kii.cloud.storage.Kii;
 import com.kii.cloud.storage.KiiBucket;
